working on create user

+++ project-app/routes/web.php
@@ -42,9 +42,12 @@ Route::middleware(['adminUserType','auth', 'verified'])->group(function(){
         return view('admin.home');
     })->name('admin.home');
 
-Route::get('/userList', [UserController::class, 'getUserList'])->name('userList');
-Route::put('/user/update', [UserController::class, 'update'])->name('user.update');
-Route::delete('/user/{id}', [UserController::class, 'delete'])->name('user.delete');
+    Route::get('/admin/user-list', [UserController::class, 'getUserList'])->name('userList');
+    Route::put('/admin/user-update', [UserController::class, 'update'])->name('user.update');
+    Route::delete('/admin/user-{id}', [UserController::class, 'delete'])->name('user.delete');
+    Route::get('/admin/user-list/search', [UserController::class, 'search'])->name('searchUsers');
+    Route::get('/admin/user-create', [UserController::class, 'create'])->name('users.create');
+    Route::post('/admin/users', [UserController::class, 'store'])->name('users.store');
 
 });
